Add new classes for programming assignment 08
The commit introduces a variety of new classes needed for programming assignment. This includes PointMain, ArgsMain, Vector2D, Vector, VectorMain, Point3D and Point2D. These classes provide the basic structure for playing with and operating on 2D and 3D vectors and points. In addition, ArgsMain is a basic program for handling user inputs.